requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
httpx==0.26.0
apscheduler==3.10.4
anthropic==0.18.1
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0
feedparser==6.0.11
aiohttp==3.9.1
tenacity==8.2.3
